+++ libraries/base/src/Data/Bounded.hs
@@ -20,3 +20,8 @@ module Data.Bounded
 
 import GHC.Enum
 
+-- | A list of all elements between 'minBound' and 'maxBound', inclusively.
+--
+-- @since 4.19.0.0
+enumerate :: (Enum a, Bounded a) => [a]
+enumerate = [minBound .. maxBound]